requirements.txt
streamlit==1.45.0
opencv-python==4.9.0.80
numpy==1.26.4
easyocr==1.7.2
torch==2.2.0
torchvision==0.17.0
ultralytics==8.3.131
pandas==2.2.3
Pillow==10.2.0
scikit-image==0.25.2
requests==2.32.3
python-dateutil==2.9.0.post0
tqdm==4.67.1
matplotlib==3.8.3
regex==2023.12.25
pytesseract==0.3.10
imutils==0.5.4 